02.echo-bot/requirements.txt
aiohttp

botbuilder-core
botbuilder-ai
flask